+++ app/models/draw_students_update.rb
@@ -5,9 +5,10 @@ class DrawStudentsUpdate
   include ActiveModel::Model
   include Callable
 
-  attr_reader :class_year, :to_add
+  attr_reader :class_year
 
   # Initialize a new DrawStudentsUpdate
+  #
   #
   # @param draw [Draw] the draw to be updated
   # @param params [#to_h] the parameters from the form
@@ -43,10 +44,8 @@ class DrawStudentsUpdate
   def process_params(params)
     @params = params.to_h.transform_keys(&:to_sym)
     @params[:class_year] = nil if @params[:class_year]&.empty?
-    @params[:to_add] = nil if @params[:to_add]&.empty?
-    @to_add = @params[:to_add]
     @class_year = set_class_year
-    @students_to_add = find_students_to_add
+    @students_to_add = students_to_add_by_class_year
   end
 
   def set_class_year
@@ -54,23 +53,11 @@ class DrawStudentsUpdate
     params[:class_year].to_i
   end
 
-  def find_students_to_add
-    students_to_add_by_class_year + students_to_add_by_username
-  end
-
   def students_to_add_by_class_year
     return [] unless class_year
     UngroupedStudentsQuery.call.includes(:draw_membership)
                           .where(class_year: class_year,
                                  draw_memberships: { draw_id: nil })
-  end
-
-  def students_to_add_by_username
-    return [] unless to_add
-    user = UngroupedStudentsQuery.call.find_by(username: to_add)
-    return [] unless user
-    return [] unless (user.student? || user.rep?) && user.membership.nil?
-    [user.remove_draw]
   end
 
   def no_action_warning